squad-goals - Refactor squad goals operator, move more code into modules, organize directory
