+++ .gitignore
@@ -2,7 +2,7 @@
 .idea
 logfile.txt
 pid.txt
-errors.log
+logs/errors.log
 .byebug_history
 run.pid
 credentials.json
