+++ 420bot.sh
@@ -1,4 +1,3 @@
 #!/bin/sh
 
 nohup bundle exec ruby 420bot.rb > logfile.txt 2>&1 & echo $! > run.pid
-
