--- config.yaml
@@ -1,2 +1,0 @@
-channel: '#dev_private'
-time_zone: 'America/New_York'
